requirements.txt
Flask
pillow
opencv-python
pandas
matplotlib
numpy
unittest